+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2025-01-29
+## Added
+- WireGuard plugin.
+- IPSec plugin.
+
+## Fixed
+- Several CI fixes and actions updates.
+- Changes needed for upgrade to the latest `validator` crate.
+- Updated Rust packages to the latest version.
+
+
 ## [2.0.1] - 2024-04-21
 ## Added
 - Crates update to the latest supported versions.